requirements.txt
anyio==4.4.0
cffi==1.17.1
click==8.1.7
cryptography==43.0.1
exceptiongroup==1.2.2
fastapi==0.97.0
h11==0.14.0
idna==3.9
pycparser==2.22
pydantic==1.10.18
PyMySQL==1.1.1
python-dotenv==1.0.1
sniffio==1.3.1
SQLAlchemy==2.0.34
starlette==0.27.0
typing_extensions==4.12.2
uvicorn==0.30.6
